nearly done

+++ BubblePop/MathUtils.swift
@@ -12,13 +12,13 @@ import UIKit
 //Purely static enum. Cannot be initialised, only for use as a static
 //collection of utility methods.
 enum MathUtils {
-    static func Distance(_ from: CGPoint, _ to: CGPoint) -> Float {
-        return self.Distance((x: Float(from.x), y: Float(from.y)), (x: Float(to.x), y: Float(to.y)))
+    static func Distance(_ from: CGPoint, _ to: CGPoint) -> Double {
+        return self.Distance((x: Double(from.x), y: Double(from.y)), (x: Double(to.x), y: Double(to.y)))
     }
     
-    static func Distance(_ from: (x: Float, y: Float), _ to: (x: Float, y: Float)) -> Float {
+    static func Distance(_ from: (x: Double, y: Double), _ to: (x: Double, y: Double)) -> Double {
         let diffX = to.x - from.x
         let diffY = to.y - from.y
-        return sqrt(powf(diffX, 2) + powf(diffY, 2))
+        return sqrt(pow(diffX, 2) + pow(diffY, 2))
     }
 }
